Transfer verification_result from verifier to core
